Project with HATEOAS done

+++ exemplo-maven-parent/src/main/java/br/senaigo/mobile/interfaces/GenericOperationsController.java
@@ -12,6 +12,6 @@ public interface GenericOperationsController<E> {
     void put(E entity); // PUT 204 - NO CONTENT
     void delete(E entity); // DELETE 204
     Resources<E> get(); // GET 200
-    Resource<E> get(Integer id); //GET 200
+    Resource<E> get(Long id); //GET 200
     void patch(E entity); //PATCH 204
 }
